requirements/base.txt
boto3>=1.26
fastapi>=0.99
pydantic>=1.10
requests>=2.31
sqlalchemy>=2.0
alembic>=1.11
sentry-sdk>=1
uvicorn[standard]>=0.22
psycopg2-binary>=2.9
python-keycloak>=3.3